requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
backtest==0.7.8
backtesting==0.6.4
backtrader==1.9.78.123
beartype==0.21.0
beautifulsoup4==4.13.4
blis==1.3.0
bokeh==3.7.3
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
curl_cffi==0.11.1
cycler==0.12.1
cymem==2.0.11
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
dm-tree==0.1.9
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
executing==2.2.0
fastai==2.7.19
fastcore==1.7.29
fastdownload==0.0.7
fastprogress==1.0.3
fasttransform==0.0.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
h5py==3.13.0
ibapi==9.81.1.post1
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keras==3.9.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.0.9
narwhals==1.40.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pandas-ta==0.3.14b0
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==5.24.1
plum-dispatch==2.5.7
preshed==3.0.9
promise==2.3
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyts==0.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-datasets==4.9.8
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.1
termcolor==3.1.0
tf_keras==2.17.0
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tsai==0.4.0
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vectorbt==0.27.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
wurlitzer==3.1.1
xyzservices==2025.4.0
ydf==0.11.0
yfinance==0.2.61
zipp==3.21.0
